feat(DST-1136): Make clear that `"stretch"` is the default for `<Stack>`
Revert making `left` the default of Stack. Might cause porblems because we would need to make `<Inline>` and `<Tiles>` to take full width and this would cause nested `<Inline>` to stop working.

By showing that a `<Stack>` will stretch by default, it is hopefully easier to reason about why it stretches things.

Note that this PR does not "fix" the `<Button>` it WILL stretch in a `<Stack>`. Having this exception might be weird. So either set the alignment to left or wrap the button in an extra element.